baseurl dan notifikasi berhasil

+++ app/src/main/java/com/dev/firedetector/util/DangerConditions.kt
@@ -7,7 +7,7 @@ data class DangerConditions(
 ) {
     companion object {
         fun fromSensorData(
-            temperature: Float,
+            temperature: Double,
             mqStatus: String,
             flameStatus: String
         ): DangerConditions {
